Moving all test utilities in correct package
