+++ core/src/test/groovy/com/novoda/gradle/release/GeneratePomFileForMavenPublicationTest.groovy
@@ -1,7 +1,7 @@
 package com.novoda.gradle.release
 
-import com.novoda.gradle.release.test.GradleScriptTemplates
-import com.novoda.gradle.release.test.TestProjectRule
+import com.novoda.gradle.test.GradleScriptTemplates
+import com.novoda.gradle.test.TestProjectRule
 import com.novoda.gradle.truth.GradleTruth
 import org.junit.Rule
 import org.junit.Test
